# [wasm] More typescript conversion