--- a/src/mono/wasm/runtime/types.ts
+++ b/src/mono/wasm/runtime/types.ts
@@ -3,7 +3,6 @@
 
 export type GCHandle = {
     __brand: "GCHandle"
-
 }
 export type JSHandle = {
     __brand: "JSHandle"