… src/mono/wasm/runtime/CMakeLists.txt
… src/mono/wasm/runtime/base64.ts
… src/mono/wasm/runtime/binding_support.js
… src/mono/wasm/runtime/dotnet_support.js
… src/mono/wasm/runtime/event-listener.ts
… src/mono/wasm/runtime/exports.ts
… src/mono/wasm/runtime/gc-handles.ts
… src/mono/wasm/runtime/icu.ts
… src/mono/wasm/runtime/js-to-cs.ts
… src/mono/wasm/runtime/library-dotnet.js
… src/mono/wasm/runtime/library_mono.js
… src/mono/wasm/runtime/method-binding.ts
… src/mono/wasm/runtime/method-calls.ts
… src/mono/wasm/runtime/modules.ts
… src/mono/wasm/runtime/profiler.ts
… src/mono/wasm/runtime/queue.ts
… src/mono/wasm/runtime/rollup.config.js
… src/mono/wasm/runtime/roots.ts
… src/mono/wasm/runtime/scheduling.ts
… src/mono/wasm/runtime/src/binding/types.ts
… src/mono/wasm/runtime/src/mono/cwraps.ts
… src/mono/wasm/runtime/src/mono/string-decoder.ts
… src/mono/wasm/runtime/src/mono/types.ts
… src/mono/wasm/runtime/src/startup.ts
… src/mono/wasm/runtime/startup.ts
… src/mono/wasm/runtime/strings.ts
… src/mono/wasm/runtime/types/emscripten.d.ts
… src/mono/wasm/runtime/types/v8.d.ts
… src/mono/wasm/runtime/web-socket.ts
… src/mono/wasm/wasm.proj