- flattened typescript folders
- original `binding_support.js` and `binding_support.js` split into multiple TS files
- clarified exports for MONO, BINDING and linking
- linking js to C methods is generated during emcc module loading, which is evaluated during compilation
- only one emscripten library file `library-dotnet.js`