@pavelsavara 2021-09-30
The new dotnet.xxx.js is 45099 bytes smaller :-)

![image](https://user-images.githubusercontent.com/271576/135451126-0cd899d6-415f-4b75-b4c4-f8b085e80d25.png)