@pavelsavara 2021-09-30
Perf test: `src\mono\sample\wasm\browser-bench `

| measurement | old time | new time |
|-:|-:|-:|
|        Exceptions, NoExceptionHandling |     0.2106us |     0.2094us |
|                   Exceptions, TryCatch |     0.2246us |     0.2220us |
|              Exceptions, TryCatchThrow |     0.0083ms |     0.0086ms |
|             Exceptions, TryCatchFilter |     0.1729us |     0.1701us |
|       Exceptions, TryCatchFilterInline |     0.0326us |     0.0314us |
|        Exceptions, TryCatchFilterThrow |     0.0095ms |     0.0098ms |
| Exceptions, TryCatchFilterThrowApplies |     0.0089ms |     0.0090ms |
|         Json, non-ASCII text serialize |     0.5308ms |     0.5361ms |
|       Json, non-ASCII text deserialize |     1.5531ms |     1.5450ms |
|                  Json, small serialize |     0.1226ms |     0.1171ms |
|                Json, small deserialize |     0.1979ms |     0.1967ms |
|                  Json, large serialize |    32.2857ms |    31.5786ms |
|                Json, large deserialize |    51.9556ms |    53.4719ms |
|              WebSocket, PartialSend 1B |     0.0016ms |     0.0018ms |
|            WebSocket, PartialSend 64KB |     0.1057ms |     0.1035ms |
|             WebSocket, PartialSend 1MB |     1.6000ms |     1.8636ms |
|           WebSocket, PartialReceive 1B |     0.0031ms |     0.0031ms |
|         WebSocket, PartialReceive 10KB |     0.0080ms |     0.0100ms |
|        WebSocket, PartialReceive 100KB |     0.0100ms |     0.0100ms |

My conclusion - no change.