@pavelsavara 2021-10-04
There is unrelated stack overflow in System.Runtime.Serialization.Xml.Tests see https://github.com/dotnet/runtime/issues/59926